minor refactoring of test-env-js directories
