+++ packages/js/test-env/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./await";
+export * from "./generate-name";
